+++ include/lttng/events.h
@@ -499,7 +499,7 @@ struct lttng_channel_ops {
 
 struct lttng_counter_ops {
 	struct lib_counter *(*counter_create)(size_t nr_dimensions,
-			const size_t *max_nr_elem,	/* for each dimension */
+			const int64_t *max_nr_elem,	/* for each dimension */
 			int64_t global_sum_step);
 	void (*counter_destroy)(struct lib_counter *counter);
 	void (*counter_add)(struct lib_counter *counter, int64_t *dimension_indexes,
